ErrorHandlingCRUDs

+++ WebshopAdmin/WebshopAdmin/MainWindow.xaml.cs
@@ -26,11 +26,11 @@ namespace WebshopAdmin
         private ListBox boxProducts = new ListBox();
         private Service.Service service;
         private Image img = new Image();
-        
+
         lewebshopEntities db = Dao.Database.db;
         public MainWindow()
         {
-           
+
             service = new Service.Service();
 
             grid.ItemsSource = service.getProducts();
@@ -61,7 +61,7 @@ namespace WebshopAdmin
             grid.Width = 275;
             grid.Height = 200;
             grid.SelectionChanged += new SelectionChangedEventHandler(Event_SelectionChanged);
-            
+
             sp_view.Children.Add(grid);
 
 
@@ -100,7 +100,7 @@ namespace WebshopAdmin
             {
                 grid.SelectedItem = l.ElementAt(0);
             }
-            
+
         }
 
         //Denne metode sætter package
@@ -111,7 +111,7 @@ namespace WebshopAdmin
 
             boxProducts.Width = 200;
             boxProducts.Height = 100;
-            boxProducts.DataContext = service.getProducts();
+            boxProducts.ItemsSource = service.getProducts();
             boxProducts.SelectionMode = SelectionMode.Multiple;
             sp_middle.Children.Add(boxProducts);
 
@@ -268,9 +268,18 @@ namespace WebshopAdmin
                             l.Add((Product)i);
                         }
 
-                        service.createPackage(l, textBox.Text, Convert.ToDecimal(textBox1.Text));
-                        gridPackage.ItemsSource = null;
-                        gridPackage.ItemsSource = db.Packages.ToList();
+
+                        decimal d;
+                        if (decimal.TryParse(textBox1.Text, out d) && textBox.Text != "")
+                        {
+                            service.createPackage(l, textBox.Text, Convert.ToDecimal(textBox1.Text));
+                            gridPackage.ItemsSource = null;
+                            gridPackage.ItemsSource = db.Packages.ToList();
+                        }
+                        else
+                        {
+                            MessageBox.Show("Name must be letters and price must be a number");
+                        }
 
                         break;
 
@@ -278,35 +287,49 @@ namespace WebshopAdmin
                         {
                             var datatable = sp_view.Children.OfType<DataGrid>().First();
                             Package p = (Package)datatable.SelectedItem;
-                            service.deletePackage(p);
-                            gridPackage.ItemsSource = null;
-                            gridPackage.ItemsSource = db.Packages.ToList();
+                            if (p != null)
+                            {
+                                service.deletePackage(p);
+                                gridPackage.ItemsSource = null;
+                                gridPackage.ItemsSource = db.Packages.ToList();
+                            }
+                            else
+                            {
+                                MessageBox.Show("Select a package");
+                            }
                         }
                         break;
                     case "Edit":
-                        if (gridPackage.SelectedItem != null)
                         {
-                            var datatable = sp_view.Children.OfType<DataGrid>().First();
-                            var textboxname = sp_middle.Children.OfType<TextBox>().First();
-                            var textboxprice = sp_middle.Children.OfType<TextBox>().ElementAt(1);
-                            var listboxproducts = sp_middle.Children.OfType<ListBox>().First();
-                            List<Product> list = new List<Product>();
-                            Package p = (Package)datatable.SelectedItem;
-                            listboxproducts.SelectionMode = SelectionMode.Multiple;
-                            foreach (var select in listboxproducts.SelectedItems)
-                            {
-                                list.Add((Product)select);
-                            }
+                            
+                                var datatable = sp_view.Children.OfType<DataGrid>().First();
+                                var textboxname = sp_middle.Children.OfType<TextBox>().First();
+                                var textboxprice = sp_middle.Children.OfType<TextBox>().ElementAt(1);
+                                var listboxproducts = sp_middle.Children.OfType<ListBox>().First();
+                                List<Product> list = new List<Product>();
 
-                            service.editpackage(list, p, textboxname.Text, Convert.ToDecimal(textboxprice.Text));
-                            gridPackage.ItemsSource = null;
-                            gridPackage.ItemsSource = db.Packages.ToList();
+                                if (gridPackage.SelectedItem != null && decimal.TryParse(textboxprice.Text, out d) && !textboxname.Text.Equals("") && listboxproducts.SelectedItems != null)
+                                {
+                                    
+                                    Package p = (Package)datatable.SelectedItem;
+                                    listboxproducts.SelectionMode = SelectionMode.Multiple;
+                                    foreach (var select in listboxproducts.SelectedItems)
+                                    {
+                                        list.Add((Product)select);
+                                    }
+
+                                    service.editpackage(list, p, textboxname.Text, Convert.ToDecimal(textboxprice.Text));
+                                    gridPackage.ItemsSource = null;
+                                    gridPackage.ItemsSource = db.Packages.ToList();
+                                }
+
+                                else
+                                {
+                                    MessageBox.Show("Pick a row in the grid and insert name, price and pick the products");
+                                }
+                            
                         }
-                        else
-                        {
-                            MessageBox.Show("Vælg venligst en række fra datagrid");
-                        }
-                        break;
+                    break;
                 }
             }
         }
@@ -323,19 +346,31 @@ namespace WebshopAdmin
                     case "Create":
                         var textQuestion = sp_middle.Children.OfType<TextBox>().First();
                         var textAnswer = sp_middle.Children.OfType<TextBox>().ElementAt(1);
-                        service.createFAQ(textQuestion.Text, textAnswer.Text);
-                        gridFAQ.ItemsSource = null;
-                        gridFAQ.ItemsSource = db.FAQs.ToList();
-
-
+                        if (textAnswer.Text != "" && textQuestion.Text != "")
+                        {
+                            service.createFAQ(textQuestion.Text, textAnswer.Text);
+                            gridFAQ.ItemsSource = null;
+                            gridFAQ.ItemsSource = db.FAQs.ToList();
+                        }
+                        else
+                        {
+                            MessageBox.Show("Fill answer and question");
+                        }
                         break;
 
                     case "Delete":
                         {
-                            service.deleteFAQ((FAQ)gridFAQ.SelectedItem);
-                            gridFAQ.ItemsSource = null;
-                            gridFAQ.ItemsSource = db.FAQs.ToList();
-
+                            
+                            if (gridFAQ.SelectedItem != null)
+                            {
+                                service.deleteFAQ((FAQ)gridFAQ.SelectedItem);
+                                gridFAQ.ItemsSource = null;
+                                gridFAQ.ItemsSource = db.FAQs.ToList();
+                            }
+                            else
+                            {
+                                MessageBox.Show("Pick a FAQ");
+                            }
                         }
                         break;
 
@@ -345,15 +380,17 @@ namespace WebshopAdmin
                             var datatable = sp_view.Children.OfType<DataGrid>().First();
                             var textboxQuestion = sp_middle.Children.OfType<TextBox>().First();
                             var textboxAnswer = sp_middle.Children.OfType<TextBox>().ElementAt(1);
-
                             FAQ faq = (FAQ)datatable.SelectedItem;
-                            service.editFAQ(faq, textboxQuestion.Text, textboxAnswer.Text);
-                            gridFAQ.ItemsSource = null;
-                            gridFAQ.ItemsSource = db.FAQs.ToList();
-                        }
-                        else
-                        {
-                            MessageBox.Show("Vælg venligst en række fra datagrid");
+                            if (textboxQuestion.Text != "" && textboxAnswer.Text != "" && faq != null)
+                            {
+                                service.editFAQ(faq, textboxQuestion.Text, textboxAnswer.Text);
+                                gridFAQ.ItemsSource = null;
+                                gridFAQ.ItemsSource = db.FAQs.ToList();
+                            }
+                            else
+                            {
+                                MessageBox.Show("Fill answer, question and pick a FAQ");
+                            }
                         }
                         break;
                 }
